fixed cuberevo-dvb-tools updating (thanks skaman)

+++ recipes/cuberevo/cuberevo-dvb-tools.bb
@@ -45,7 +45,7 @@ RDEPENDS_append_cuberevo-9500hd = " module-init-tools-depmod"
 RDEPENDS_append_cuberevo-mini = " module-init-tools-depmod"
 RDEPENDS_append_cuberevo-mini2 = " module-init-tools-depmod"
 
-PR = "r6"
+PR = "r7"
 GCC ?= ""
 
 SRC_URI = "http://enigma2.sifteam.eu/tmp/opencuberevo/cuberevo-dvb-tools-${MACHINE}-${PV}${GCC}.tar.bz2 \
@@ -88,11 +88,11 @@ do_install() {
         install -m 0755 ${WORKDIR}/libdvbsi++.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/usr/lib/libdvbsi++.so.0.0.0
         install -m 0755 ${WORKDIR}/libdreamdvd.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/usr/lib/libdreamdvd.so.0.0.0
         install -m 0755 ${WORKDIR}/libeplayer3.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so.0.0.0
-        ln -s libeplayer3.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so.0
-        ln -s libeplayer3.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so
+        [ ! -f "${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so.0" ] && ln -s libeplayer3.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so.0
+        [ ! -f "${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so" ] && ln -s libeplayer3.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libeplayer3.so
         install -m 0755 ${WORKDIR}/libmmeimage.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so.0.0.0
-        ln -s libmmeimage.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so.0
-        ln -s libmmeimage.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so
+        [ ! -f "${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so.0" ] && ln -s libmmeimage.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so.0
+        [ ! -f "${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so" ] && ln -s libmmeimage.so.0.0.0 ${STAGING_DIR}/${HOST_SYS}/lib/libmmeimage.so
         install -m 0755 ${WORKDIR}/libeplayer3.so.0.0.0 ${D}/lib/libeplayer3.so.0.0.0
         ln -s libeplayer3.so.0.0.0 ${D}/lib/libeplayer3.so.0
         ln -s libeplayer3.so.0.0.0 ${D}/lib/libeplayer3.so
